+++ History.md
@@ -4,6 +4,7 @@ unreleased
   * "dev" format will use same tokens as other formats
   * `:response-time` token is now monotonic
   * `:response-time` token has precision to 1 μs
+  * fix `:status` + immediate output in node.js 0.8
   * deps: bytes@1.0.0
     - add negative support